+++ framework/Base/Component.php
@@ -8,9 +8,6 @@ namespace Rid\Base;
 abstract class Component extends BaseObject implements  ComponentInterface
 {
 
-    // 协程模式
-    private $_coroutineMode = ComponentInterface::COROUTINE_MODE_NEW;
-
     // 状态
     private $_status = ComponentInterface::STATUS_READY;
 
@@ -24,18 +21,6 @@ abstract class Component extends BaseObject implements  ComponentInterface
     public function setStatus($status)
     {
         $this->_status = $status;
-    }
-
-    // 获取协程模式
-    public function getCoroutineMode()
-    {
-        return $this->_coroutineMode;
-    }
-
-    // 设置协程模式
-    public function setCoroutineMode($coroutineMode)
-    {
-        $this->_coroutineMode = $coroutineMode;
     }
 
     // 请求前置事件